1. 合并cppjieba最新代码https://github.com/yanyiwu/cppjieba 2. 添加nodejs调用textrank接口，并可以指定allowedPOS 3. 对于tfidf关键词抽取接口extract，可以指定allowedPOS 4. 对于textrank和tfidf接口，增加传入分词结果，直接获取结果的接口，避免重复分词

+++ deps/cppjieba/FullSegment.hpp
@@ -48,17 +48,17 @@ class FullSegment: public SegmentBase {
   void Cut(RuneStrArray::const_iterator begin, 
         RuneStrArray::const_iterator end, 
         vector<WordRange>& res) const {
-    //resut of searching in trie tree
+    // resut of searching in trie tree
     LocalVector<pair<size_t, const DictUnit*> > tRes;
 
-    //max index of res's words
-    int maxIdx = 0;
+    // max index of res's words
+    size_t maxIdx = 0;
 
     // always equals to (uItr - begin)
-    int uIdx = 0;
+    size_t uIdx = 0;
 
-    //tmp variables
-    int wordLen = 0;
+    // tmp variables
+    size_t wordLen = 0;
     assert(dictTrie_);
     vector<struct Dag> dags;
     dictTrie_->Find(begin, end, dags);
